SCUFFED: fixed slowdown when making newline in large file
This involved doing some incredibly suspicious activity, and I'm just going to let it be for now so I can focus on other features. In the future if I can think of a way to make it less scuffed I will. If this gets merged into main before I can think of something then I'll detail the scuffed stuff on the PR page